+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change log
 
 ## [Unreleased]
+### Added
+- data_entropy()
+
+### Changed
+- histogram(): remove unique elements of array as optional output
 
 ## [0.7.1] - 2018-06-22
 ### Changed